+++ RevitLookup/Extension/ObjectExtension.cs
@@ -54,6 +54,8 @@ namespace RevitLookupWpf.Extension
             //Methods
             var methodInfos = type.GetMethods()
                 .Where(p => !p.Name.StartsWith("get_") && !p.Name.StartsWith("set_"))
+                .Where(m => !m.Name.StartsWith("add_") && !m.Name.StartsWith("remove_"))
+                .Where(xyz=>!xyz.Name.StartsWith("op_"))
                 .Where(p => p.Name != "Dispose")
                 .ToList();
             //.OrderBy(p => p.Name);